Remove all traces of redux
