--- app/javascript/packs/pages.js
@@ -1,1 +1,0 @@
-import Recipe from "../pages/Recipe"
