+++ app/javascript/pages/Recipe/styles.scss
@@ -1,0 +1,1 @@
+
