--- app/javascript/reducers/index.js
@@ -1,5 +1,0 @@
-import { combineReducers } from "redux";
-
-import recipes from './recipes'
-
-export default combineReducers({ recipes });
