--- app/javascript/stores/index.js
@@ -1,12 +1,0 @@
-import { createStore } from "redux";
-import rootReducer from "../reducers";
-
-// {
-//   recipes: [
-//     {},
-//     {},
-//     {}
-//   ]
-// }
-
-export default createStore(rootReducer);
